add all activity filter to ff-projects, minor refactoring
